Refactor all double prices to big decimal

+++ src/main/resources/db/migration/V3__Create_Package_Table.sql
@@ -6,7 +6,7 @@ CREATE TABLE package (
   plate_id BIGINT,
   pieces_carton INT,
   photo VARCHAR(255),
-  price DOUBLE,
+    price DECIMAL(10, 2),
   FOREIGN KEY (plate_id) REFERENCES plate(id),
   FOREIGN KEY (carton_id) REFERENCES carton(id),
   is_deleted TINYINT(1) DEFAULT FALSE
